优化mongo执行

+++ sql/engines/mongo.py
@@ -334,6 +334,7 @@ class MongoEngine(EngineBase):
                         sql=sql,
                         auth_db=auth_db,
                     )
+                logger.error("mongo执行cmd：" + cmd)
                 p = subprocess.Popen(
                     cmd,
                     shell=True,
@@ -421,10 +422,11 @@ class MongoEngine(EngineBase):
                 try:
                     # DeprecationWarning: time.clock has been deprecated in Python 3.3 and will be removed from Python 3.8: use time.perf_counter or time.process_time instead
                     start = time.perf_counter()
+                    logger.error("执行语句：" + exec_sql)
                     r = self.exec_cmd(exec_sql, db_name)
                     end = time.perf_counter()
                     line += 1
-                    logger.debug("执行结果：" + r)
+                    logger.error("执行结果1：" + r)
                     # 如果执行中有错误
                     rz = r.replace(" ", "").replace('"', "")
                     tr = 1
@@ -459,18 +461,21 @@ class MongoEngine(EngineBase):
                             methodStr = exec_sql.split(").")[-1].split("(")[0].strip()
                             if "." in methodStr:
                                 methodStr = methodStr.split(".")[-1]
-                            if methodStr == "insert":
-                                actual_affected_rows = r.get("nInserted", 0)
-                            elif methodStr in ("insertOne", "insertMany"):
-                                actual_affected_rows = r.count("ObjectId")
-                            elif methodStr == "update":
-                                actual_affected_rows = r.get("nModified", 0)
-                            elif methodStr in ("updateOne", "updateMany"):
-                                actual_affected_rows = r.get("modifiedCount", 0)
-                            elif methodStr in ("deleteOne", "deleteMany"):
-                                actual_affected_rows = r.get("deletedCount", 0)
-                            elif methodStr == "remove":
-                                actual_affected_rows = r.get("nRemoved", 0)
+                            if not isinstance(r, str):
+                                if methodStr == "insert":
+                                    actual_affected_rows = r.get("nInserted", 0)
+                                elif methodStr in ("insertOne", "insertMany"):
+                                    actual_affected_rows = r.count("ObjectId")
+                                elif methodStr == "update":
+                                    actual_affected_rows = r.get("nModified", 0)
+                                elif methodStr in ("updateOne", "updateMany"):
+                                    actual_affected_rows = r.get("modifiedCount", 0)
+                                elif methodStr in ("deleteOne", "deleteMany"):
+                                    actual_affected_rows = r.get("deletedCount", 0)
+                                elif methodStr == "remove":
+                                    actual_affected_rows = r.get("nRemoved", 0)
+                                else:
+                                    actual_affected_rows = 0
                             else:
                                 actual_affected_rows = 0
                         # 把结果转换为ReviewSet
